@@ -2,24 +2,8 @@
 
 namespace Baumeister.Examples.ExampleAggregate
 {
+    [Builder(typeof(Address))]
     public class AddressBuilder : BuilderBase<AddressBuilder, Address>
     {
-        public AddressBuilder WithStreet(Street street)
-        {
-            With(street);
-            return this;
-        }
-
-        public AddressBuilder WithCity(City city)
-        {
-            With(city);
-            return this;
-        }
-
-        public AddressBuilder WithZipCode(ZipCode zipCode)
-        {
-            With(zipCode);
-            return this;
-        }
     }
 }
